TopologicalGiftWrapping8A.frag_face_channel — Function

frag_face_channel(in_chan, out_chan, V, EV, FE, sp_idx)

Parallel fragmentation of faces in FE against faces in sp_idx.

TopologicalGiftWrapping8A.frag_face — Function

frag_face(V::Points, copEV::ChainOp, copFE::ChainOp; [multiproc::Bool])

Performs the 2D transformation of the faces provided as input via the sigma parameter and finally each face in this sigma parameter intersects with the faces in sp index

TopologicalGiftWrapping8A.merge_vertices — Function

merge_vertices(V::Lar.Points, EV::Lar.ChainOp, FE::Lar.ChainOp, err::Float64)

Performs merge of neighboring faces, vertices and sides

Additional arguments:

err::Float64: Hyperparameter. Defaults to 1e-4.

TopologicalGiftWrapping8A.spatial_arrangement_1 — Function

spatial_arrangement_1(V::Points, copEV::ChainOp, copFE::ChainOp; [multiproc::Bool])

First step of spatial arrangement. It computes and fragments the faces in order to recall planar arrangement.

Additional arguments:

multiproc::Bool: Runs the computation in parallel mode. Defaults to false.

TopologicalGiftWrapping8A.removeinnerloops — Function

removeinnerloops(g, nFE)

Remove the faces within the inner loops from sparse matrix nFE. The return walue has g less rows than the input nFE.

TopologicalGiftWrapping8A.spatial_arrangement — Function

spatial_arrangement(V::Points, copEV::ChainOp, copFE::ChainOp; [multiproc::Bool])

Compute the arrangement on the given cellular complex 2-skeleton in 3D.

A cellular complex is arranged when the intersection of every possible pair of cell of the complex is empty and the union of all the cells is the whole Euclidean space. The function returns the full arranged complex as a list of vertices V and a chain of borders EV, FE, CF.

Additional arguments:

multiproc::Bool: Runs the computation in parallel mode. Defaults to false.